Refactored TemporalMapper to use the new Mapper class
