+++ src/temporalmapper/__init__.py
@@ -2,9 +2,8 @@ import temporalmapper.temporal_mapper as tm_main
 import temporalmapper.plotting as plot
 import temporalmapper.kernels as kernels
 import temporalmapper.cluster as cluster
-import temporalmapper.mapper as mapper
 
-Mapper = mapper.Mapper
+Mapper = tm_main.Mapper
 TemporalMapper = tm_main.TemporalMapper
 centroid_datamap = plot.centroid_datamap
 time_semantic_plot = plot.time_semantic_plot
