+++ src/temporalmapper/temporal_mapper.py
@@ -13,8 +13,8 @@ from scipy.sparse import issparse
 from sklearn.preprocessing import StandardScaler
 from sklearn.decomposition import PCA
 from sklearn.neighbors import NearestNeighbors
-from sklearn.base import BaseEstimator, ClusterMixin
-from sklearn.utils.validation import check_is_fitted
+from sklearn.base import BaseEstimator, ClusterMixin, clone
+from sklearn.utils.validation import check_is_fitted, check_array
 
 from temporalmapper.utilities import(
     std_sigmoid,
@@ -24,6 +24,7 @@ from temporalmapper.utilities import(
 from temporalmapper.plotting import (
     time_semantic_plot,
 )
+
 from temporalmapper.layout import compute_time_semantic_positions
 from temporalmapper.kernels import square
 from temporalmapper.analytics import (
@@ -36,7 +37,7 @@ minimal usage example:
     # load from your data file:
     data : (n_dim, N_data) array-like
     time : (N_data,) array-like
-    # choose an sklearn clusterer:
+    # choose an sklearn-compliant clusterer:
     clusterer = HDBSCAN()
 
     # init and build the graph:
@@ -85,11 +86,10 @@ class TemporalMapper(BaseEstimator):
         time: npt.NDArray | None=None, # backwards
         data: npt.NDArray | None=None, # compatibility
         clusterer: ClusterMixin=None,
-        N_checkpoints: int=None,
+        N_checkpoints: int=5,
         neighbours: int=5,
         overlap: float=0.5,
         inclusion_threshold: float=0.01,
-        checkpoints: list[float]=None,
         show_outliers: bool=False,
         slice_method: str="time",
         rate_sensitivity: int=1,
@@ -142,328 +142,103 @@ class TemporalMapper(BaseEstimator):
                 data = data.reshape(-1, 1)
             self.data = np.array(data)
             
-        self.checkpoints = checkpoints
         if slice_method in ["time", "data"]:
             self.slice_method = slice_method
         else:
             raise AttributeError("Accepted slice_method is 'time' or 'data'.")
-        if checkpoints is not None:
-            self.N_checkpoints = np.size(checkpoints)
-            if N_checkpoints is not None and (
-                not (self.N_checkpoints == N_checkpoints)
-            ):
-                raise AttributeError(
-                    "Given checkpoints and N_checkpoints, len(checkpoints) must equal N_checkpoints."
-                )
-        else:
-            if N_checkpoints is not None:
-                self.N_checkpoints = N_checkpoints
-            else:
-                raise AttributeError(
-                    "You must pass one of checkpoints or N_checkpoints."
-                )
-
         self.clusterer = clusterer
-        self.clusters = None
+        self.N_checkpoints = N_checkpoints
         self.inclusion_threshold = inclusion_threshold
         self.overlap = overlap
-        self.g = self.overlap
-        self.density = None
         self.rate = None
         self.rate_sensitivity = rate_sensitivity
         self.sensitivity = self.rate_sensitivity
         self.kernel = kernel
         self.kernel_params = kernel_params
-        self.G = nx.DiGraph()
-        self.adj_matrix = None
         self.pos = None
         self.verbose = verbose
         self.disable = not verbose  # for tqdm
         self.show_outliers = False
         self.neighbours = neighbours 
-        self.k = self.neighbours 
-        self.distance = None
-        self.cbeta = None
-
-    def _compute_checkpoints(self):
-        """Compute evenly spaced checkpoints at which to center the mapper slices."""
-        if self.slice_method == "data":
-            idx = np.linspace(0, self.n_samples, self.N_checkpoints + 2)[1:-1]
-            idx = np.array([int(x) for x in idx])
-            checkpoints = self.time[idx]
-        if self.slice_method == "time":
-            checkpoints = np.linspace(
-                np.amin(self.time), np.amax(self.time), self.N_checkpoints + 2
-            )[1:-1]
-        self.checkpoints = checkpoints
-
-        return checkpoints
-
-    def _compute_knn(self):
-        """Run sklearn NearestNeighbours to compute knns."""
-        if not isinstance(self.k, int) or self.k <= 0:
-            raise ValueError("'k' must be a positive integer.")
-        if self.k > self.n_samples:
-            raise ValueError(
-                f"'k' (neighbours={self.k}) must be <= n_samples = "
-                f"{self.n_samples}."
-            )
-        if self.verbose:
-            print("Computing k nearest neighbours...")
-        std_time = np.copy(self.time)
-        std_time = self.scaler.fit_transform(std_time.reshape(-1, 1))
-        datatime = np.concatenate((self.data, std_time), axis=1)
-        nbrs = NearestNeighbors(n_neighbors=self.k).fit(datatime)
-        self.distance, self.dist_indices = nbrs.kneighbors(datatime)
-        return self.distance, self.dist_indices
-
-    def _compute_density(self):
-        """Compute the temporal density (f-rate) at each point."""
-        if self.sensitivity == 0:
-            if self.verbose:
-                print(
-                    "Temporal density sensitivity is set to 0, skipping density computation."
-                )
-            self.density = np.ones(self.n_samples)
-            return self.density
-        if self.distance is None:
-            self._compute_knn()
-        if self.verbose:
-            print("Computing spatial density...")
-        self.data_width = np.mean(
-            [
-                np.amax(self.data[:, k]) - np.amin(self.data[:, k])
-                for k in range(self.data.shape[1])
-            ]
+        density_based = False if rate_sensitivity == 0 else True
+        self._mapper = Mapper(
+            clusterer = clusterer,
+            n_slices = self.N_checkpoints,
+            n_neighbors = self.neighbours,
+            overlap = self.overlap,
+            inclusion_threshold = self.inclusion_threshold,
+            slice_method = self.slice_method,
+            density_based = density_based,
+            kernel = self.kernel,
+            kernel_params = self.kernel_params,
+            time_index = -1,
+            verbose = int(self.verbose),
         )
-        radius = self.distance[:, -1]
-        density = self.k * np.ones(self.n_samples)
-        temporal_width = np.array(
-            [max(self.time[idx]) - min(self.time[idx]) for idx in self.dist_indices]
-        )
-        density /= temporal_width
-
-        # apply the smoothing window:
-        d_window = self.data_width / 10
-        smoothed_densities = np.array(
-            [
-                np.average(
-                    density[idx], weights=cosine_window(self.distance[k], d_window)
-                )
-                for k, idx in enumerate(self.dist_indices)
-            ]
-        )
-        smoothed_densities = std_sigmoid(smoothed_densities)
-
-        if self.sensitivity == -1:
-            self.density = 1 / (1 - np.log2(smoothed_densities))
-        else:
-            self.density = smoothed_densities**self.sensitivity
-        return self.density
-
-    def _compute_kernel_width(self):
-        """Return the parameter c(beta) for the kernel width.
-        Currently unused.
-        """
-        if self.density is None:
-            self._compute_density()
-        sorting_index = np.argsort(self.density)
-        reverse_sort_dict = {sorting_index[i]: i for i in range(self.n_samples)}
-        reverse_sort_index = np.zeros(self.n_samples, dtype=int)
-        for s in sorting_index:
-            reverse_sort_index[s] = reverse_sort_dict[s]
-        cdf = reverse_sort_index / self.n_samples
-        c_max = 2  # todo magic number
-        self.cbeta = c_max * cdf + (1 - cdf)
-        return self.cbeta
-
-    def _cluster(self):
-        """At each checkpoint, use the clustering algorithm to cluster the points in the
-        associated bin. A convention here is that a cluster of -1 means noise, and a
-        cluster of -2 means unclustered.
-        """
-        if self.clusterer is None:
-            raise ValueError("`self.clusterer is None`")
-        if self.checkpoints is None:
-            self._compute_checkpoints()
-        if self.density is None:
-            self._compute_density()
-        if self.cbeta is None:
-            self._compute_kernel_width()
-        if self.verbose:
-            print("Clustering at each time slice.")
-        clusters, weights = weighted_clusters(
-            self.data,
-            self.time,
-            self.checkpoints,
-            self.density / np.median(self.density),
-            self.clusterer,
-            self.kernel,
-            self.g,
-            self.kernel_params,
-            eps=self.inclusion_threshold,
-        )
-        self.clusters = clusters
-        self.weights = weights
-        if not np.all(np.any(weights > 0, axis=1)):
-            # in theory this shouldn't happen, but it does sometimes
-            # (todo)
-            print("Warning: Your mapper params do not form a cover.")
-        return clusters
-
-    def add_vertices(self, y_data=1):
-        """Add the clusters from each time slice as vertices in the networkx graph (self.G)."""
-        node_counter = 0
-        slices = []
-        for i in trange(
-            self.N_checkpoints,
-            disable=self.disable,
-            desc="Converting clusters to vertices",
-        ):
-            slice_idx = (self.clusters[i] != -2).nonzero()[0]
-            slices.append(slice_idx)
-            cluster = self.clusters[i][slice_idx]
-            for l, val in enumerate(np.unique(cluster)):
-                if (val == -1) and (not self.show_outliers):
-                    continue
-
-                # Construct the basic attributes of the node:
-                slice_no = i
-                cluster_no = val
-                node_label = str(i) + ":" + str(val)
-
-                # Add a node with the attributes
-                self.G.add_node(
-                    node_label,
-                    slice_no=slice_no,
-                    cluster_no=cluster_no,
-                    node_number=node_counter,
-                )
-                node_counter += 1
-
-        self.slices = slices
-        if self.verbose:
-            print("%d vertices added." % (np.size(self.G.nodes())))
-
-        return self
-
-    def build_adj_matrix(self):
-        verts = self.G.nodes()
-        n_verts = len(verts)
-        verts = np.array(verts)
-        adj_mat = np.zeros((n_verts, n_verts))
-        slices = self.slices
-        time_centers = np.zeros(len(slices))
-        bin_width = np.zeros(len(slices))
-        for k, slice_ in enumerate(slices):
-            if np.size(slice_)==0:
-                time_centers[k] = self.checkpoints[k]
-                bin_width[k] = 0
-            else:
-                time_centers[k] = np.median(self.time[slice_])
-                bin_width[k] = np.max(self.time[slice_]) - np.min(self.time[slice_])
-
-        for i in trange(
-            self.N_checkpoints - 1, disable=self.disable, desc="Adding edges"
-        ):
-            clust1 = self.clusters[i]
-            clust2 = self.clusters[i + 1]
-
-            for j in slices[i]:
-                c1 = clust1[j]
-                c2 = clust2[j]
-                if ((c1 == -1) or (c2 == -1)) and (not self.show_outliers):
-                    # outliers
-                    continue
-                if (c1 == -2) or (c2 == -2):
-                    # not in both slices.
-                    continue
-                c1_str = str(i) + ":" + str(int(c1))
-                c2_str = str(i + 1) + ":" + str(int(c2))
-
-                # Get the matrix indices corresponding to the two clusters
-                l = self.G.nodes()[c1_str]["node_number"]
-                k = self.G.nodes()[c2_str]["node_number"]
-
-                adj_mat[l][k] += self.kernel(
-                    time_centers[i],
-                    self.time[j],
-                    self.density[j],
-                    bin_width[i],
-                    params=self.kernel_params,
-                )
-
-        self.adj_matrix = adj_mat
-        return self
-
-    def add_edges(self):
-        if type(self.adj_matrix) != np.ndarray:
-            self.build_adj_matrix()
-        # Use the adj. matrix to add the weighted edges
-        i = j = 0
-        verts = np.array(self.G.nodes())
-        for row in self.adj_matrix:
-            i = 0
-            for val in row:
-                if val == 0:
-                    i += 1
-                    continue
-                self.G.add_edge(verts[j], verts[i], weight=val)
-                i += 1
-            j += 1
-        return self
 
     def build(self):
         """ Construct the density-based Mapper graph """
-        # quick data prep
-        self.n_samples = np.size(self.time)
-        data = self.data
-        if len(data.shape) == 1:
-            data = data.reshape(-1, 1)
-        self.n_components = data.shape[1]
-        if issparse(data):
-            self.scaler = StandardScaler(copy=False, with_mean=False)
-        else:
-            self.scaler = StandardScaler(copy=False)
-        self.data = self.scaler.fit_transform(data)
-        # dbmapper
-        if self.clusters is None:
-            self._cluster()
-        self.add_vertices()
-        self.build_adj_matrix()
-        self.add_edges()
-        # extra attributes
+        X = np.vstack([self.data.flatten(), self.time.flatten()]).T
+        self._mapper.fit(X)
+
         self.populate_node_attrs()
         self.populate_edge_attrs()
+        
         self.is_fitted_ = True
         return self
 
-    def fit(
-            self,
-            time: npt.NDArray=None,
-            X: npt.NDArray=None,
-        ):
-        """ Construct the density-based Mapper graph """
-        # Backward compatibility fallback
-        if time is None and X is None:
-            time = self.time
-            X = self.data
-    
-        if time is None or X is None:
-            raise AttributeError("`time` and `data` must be provided to fit().")
-    
-        if np.size(time) != np.shape(X)[0]:
+    def fit(self, X, y=None):
+        X = check_array(X)
+        self.n_features_in_ = X.shape[1]
+        time = X[:, -1]
+        data = np.delete(X, -1, axis=1)
+        if data.shape[1] == 0:
             raise ValueError(
-                "Number of datapoints",
-                np.shape(X)[0],
-                "does not equal number of timestamps",
-                np.size(time),
+                f"After removing last column (time),"
+                " found array with 0 feature(s). "
+                f"Input X must have at least 2 columns, 1 feature(s) + time"
             )
+        self.n_samples = X.shape[0]
+        self.n_components = data.shape[1]
+
+        self._mapper = self._mapper.fit(X)
+        
+        if issparse(data):
+            self.scaler_ = StandardScaler(copy=False, with_mean=False)
+        else:
+            self.scaler_ = StandardScaler(copy=False)
+        data = clone(self.scaler_).fit_transform(data)
+        self.data = data
         self.time = time
-        self.data = X
         
         return self.build()
 
+
+    """ A bunch of property getters for self._mapper """
+    @property
+    def G(self):
+        check_is_fitted(self._mapper, 'graph_')
+        return self._mapper.graph_
+
+    @property
+    def clusters(self):
+        check_is_fitted(self._mapper, ['labels_'])
+        return self._mapper.labels_
+
+    @property
+    def weights(self):
+        check_is_fitted(self._mapper, ['weights_'])
+        return self._mapper.weights_
+
+    @property
+    def density(self):
+        check_is_fitted(self._mapper, ['density_'])
+        return self._mapper.density_
+
+    @property
+    def checkpoints(self):
+        check_is_fitted(self._mapper, ['midpoints_'])
+        return self._mapper.midpoints_
+    
     def populate_edge_attrs(self):
         """Add src_weight and dst_weight properties to every edge."""
         drift = {}
@@ -814,3 +589,261 @@ class TemporalMapper(BaseEstimator):
         )
         fig.update_traces(marker_showscale=False)
         return fig
+
+
+class Mapper(BaseEstimator):
+    def __init__(
+        self,
+        clusterer: ClusterMixin,
+        n_slices: int = 5,
+        n_neighbors: int = 5, # american spelling :^(
+        overlap: float = 0.5,
+        inclusion_threshold: float = 0.1,
+        slice_method: str = "time",
+        density_based: bool = True,
+        kernel: Callable[[float,float,float,float],float] = square,
+        kernel_params: dict | None = None,
+        time_index: int = -1,
+        verbose: int = 0,
+    ):
+        self.clusterer = clusterer
+        self.n_slices = n_slices
+        self.n_neighbors = n_neighbors
+        self.overlap = overlap
+        self.inclusion_threshold = inclusion_threshold
+        self.slice_method = slice_method
+        self.density_based = density_based
+        self.kernel = kernel
+        self.kernel_params = kernel_params
+        self.time_index = time_index
+        self.verbose = verbose
+
+    def fit(self, X, y=None):
+        X = check_array(X)
+        if not (-X.shape[1] <= self.time_index < X.shape[1]):
+            raise ValueError("Invalid time_index for input shape")
+        self.n_features_in_ = X.shape[1]
+        time = X[:, self.time_index]
+        data = np.delete(X, self.time_index, axis=1)
+        if data.shape[1] == 0:
+            raise ValueError(
+                f"After removing time_index={self.time_index},"
+                " found array with 0 feature(s). "
+                f"Input X must have at least 2 columns, 1 feature(s) + time"
+            )
+        self.n_samples_ = X.shape[0]
+        self.n_components_ = X.shape[1]
+        if issparse(data):
+            self.scaler_ = StandardScaler(copy=False, with_mean=False)
+        else:
+            self.scaler_ = StandardScaler(copy=False)
+        data = clone(self.scaler_).fit_transform(data)
+
+        self._compute_midpoints(time)
+        self._compute_density(data, time)
+        self._cluster(data, time)
+        self.graph_ = nx.DiGraph()
+        self._add_vertices()
+        self._build_adjacency_matrix(time)
+        self._add_edges()
+        self.is_fitted_ = True
+        return self
+
+    def _compute_midpoints(self, time):
+        """Compute evenly spaced midpoints at which to center the mapper slices."""
+        n_samples = np.size(time)
+        if self.slice_method == "data":
+            idx = np.linspace(0, n_samples, self.n_slices + 2)[1:-1]
+            idx = np.array([int(x) for x in idx])
+            checkpoints = time[idx]
+        elif self.slice_method == "time":
+            checkpoints = np.linspace(
+                np.amin(time), np.amax(time), self.n_slices + 2
+            )[1:-1]
+        else:
+            raise ValueError(
+                f"Supported values of ``slice_method`` are 'time' and 'data',"
+                f" not {self.slice_method}"
+            )
+        self.midpoints_ = checkpoints
+
+    def _compute_knn(self, data, time):
+        """Run sklearn NearestNeighbours to compute knns."""
+        n_samples = np.size(time)
+        if not isinstance(self.n_neighbors, int) or self.n_neighbors <= 0:
+            raise ValueError("'k' must be a positive integer.")
+        if self.n_neighbors > n_samples:
+            raise ValueError(
+                f"'k' (neighbours={self.n_neighbors}) must be <= n_samples = "
+                f"{n_samples}."
+            )
+        if self.verbose:
+            print("Computing nearest neighbours...")
+        std_time = np.copy(time)
+        std_time = clone(self.scaler_).fit_transform(std_time.reshape(-1, 1))
+        datatime = np.concatenate((data, std_time), axis=1)
+        nbrs = NearestNeighbors(n_neighbors=self.n_neighbors).fit(datatime)
+        self.distance_, self.dist_indices_ = nbrs.kneighbors(datatime)
+        return self.distance_, self.dist_indices_
+
+    def _compute_density(self, data, time):
+        """Compute the temporal density (f-rate) at each point."""
+        n_samples = np.size(time)
+        if not self.density_based:
+            self.density_ = np.ones(n_samples)
+            return self.density_
+            
+        self._compute_knn(data, time)
+        if self.verbose:
+            print("Computing spatial density...")
+        data_width_ = np.mean(
+            [
+                np.amax(data[:, k]) - np.amin(data[:, k])
+                for k in range(data.shape[1])
+            ]
+        )
+        radius = self.distance_[:, -1]
+        density = self.n_neighbors * np.ones(n_samples)
+        temporal_width = np.array(
+            [max(time[idx]) - min(time[idx]) for idx in self.dist_indices_]
+        )
+        density /= temporal_width
+
+        # apply the smoothing window:
+        d_window = data_width_ / 10
+        smoothed_densities = np.array(
+            [
+                np.average(
+                    density[idx], weights=cosine_window(self.distance_[k], d_window)
+                )
+                for k, idx in enumerate(self.dist_indices_)
+            ]
+        )
+        self.density_ = std_sigmoid(smoothed_densities)
+        return self.density_
+    
+    def _cluster(self, data, time):
+        """For each slice, use the clustering algorithm to cluster the points in the
+        slice. A convention here is that a cluster of -1 means noise, and a
+        cluster of -2 means unclustered.
+        """
+        check_is_fitted(self, ["midpoints_", "density_"])
+        if self.verbose:
+            print("Clustering at each time slice.")
+        clusters, weights = weighted_clusters(
+            data,
+            time,
+            self.midpoints_,
+            self.density_ / np.median(self.density_),
+            clone(self.clusterer),
+            self.kernel,
+            self.overlap,
+            self.kernel_params,
+            eps=self.inclusion_threshold,
+        )
+        self.labels_ = clusters
+        self.weights_ = weights
+        if not np.all(np.any(weights > 0, axis=1)):
+            # in theory this shouldn't happen, but it does sometimes (todo)
+            warn("Your mapper params do not form a cover.")
+
+        slices = [(self.labels_[i] != -2).nonzero()[0] for i in range(self.n_slices)]
+        self.slices_ = slices
+            
+        return self.labels_
+
+    def _add_vertices(self):
+        """ 
+        Add the clusters from each time slice as
+        vertices in the networkx graph ``self.graph_``.
+        """
+        check_is_fitted(self, ["slices_","labels_"])
+        node_counter = 0
+        for i in range(self.n_slices):
+            slice_idx = self.slices_[i]
+            clusters = self.labels_[i][slice_idx]
+            for l, val in enumerate(np.unique(clusters)):
+                if (val == -1):
+                    # No vertex for noise points
+                    continue
+                # Construct the basic attributes of the node:
+                slice_no = i
+                cluster_no = val
+                node_label = str(i) + ":" + str(val)
+
+                self.graph_.add_node(
+                    node_label,
+                    slice_no=slice_no,
+                    cluster_no=cluster_no,
+                    node_number=node_counter,
+                )
+                node_counter += 1
+
+        if self.verbose:
+            print("%d vertices added." % (np.size(self.graph_.nodes())))
+
+        return self.graph_
+
+    def _build_adjacency_matrix(self, time):
+        check_is_fitted(self, ["slices_", "labels_"])
+        verts = self.graph_.nodes()
+        n_verts = len(verts)
+        verts = np.array(verts)
+        adj_mat = np.zeros((n_verts, n_verts))
+        slices = self.slices_
+        time_centers = np.zeros(len(slices))
+        bin_width = np.zeros(len(slices))
+        for k, slice_ in enumerate(slices):
+            if np.size(slice_)==0:
+                time_centers[k] = self.midpoints_[k]
+                bin_width[k] = 0
+            else:
+                time_centers[k] = np.median(time[slice_])
+                bin_width[k] = np.max(time[slice_]) - np.min(time[slice_])
+
+        for i in range(self.n_slices - 1):
+            clust1 = self.labels_[i]
+            clust2 = self.labels_[i + 1]
+
+            for j in slices[i]:
+                c1 = clust1[j]
+                c2 = clust2[j]
+                if ((c1 == -1) or (c2 == -1)):
+                    # outliers
+                    continue
+                if (c1 == -2) or (c2 == -2):
+                    # not in both slices.
+                    continue
+                c1_str = str(i) + ":" + str(int(c1))
+                c2_str = str(i + 1) + ":" + str(int(c2))
+
+                # Get the matrix indices corresponding to the two clusters
+                l = self.graph_.nodes()[c1_str]["node_number"]
+                k = self.graph_.nodes()[c2_str]["node_number"]
+
+                adj_mat[l][k] += self.kernel(
+                    time_centers[i],
+                    time[j],
+                    self.density_[j],
+                    bin_width[i],
+                    params=self.kernel_params,
+                )
+
+        self.adj_matrix_ = adj_mat
+        return self.adj_matrix_
+  
+    def _add_edges(self):
+        """ Use the adj. matrix to add the weighted edges """
+        check_is_fitted(self, "adj_matrix_")
+        i = j = 0
+        verts = np.array(self.graph_.nodes())
+        for row in self.adj_matrix_:
+            i = 0
+            for val in row:
+                if val == 0:
+                    i += 1
+                    continue
+                self.graph_.add_edge(verts[j], verts[i], weight=val)
+                i += 1
+            j += 1
+        return self.graph_
